requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
django-filter
psycopg[binary]
python-decouple
celery
redis
pillow
requests
openai
anthropic
python-dotenv
gunicorn
cryptography
qrcode
whitenoise
